+++ JeffFerguson.Gepsio/LinkbaseDocument.cs
@@ -33,6 +33,7 @@ namespace JeffFerguson.Gepsio
         // Used for when a linkbaseRef element has no "role" attribute and the type of linkbase
         // document must be discovered. The 331-equivalentRelationships-instance-02.xml
         // document in the XBRL-CONF-2014-12-10 conformance suite is an example of this need.
+        // A NULL reference will be returned if the correct type cannot be identified.
         //------------------------------------------------------------------------------------
         internal static LinkbaseDocument Create(string containingDocumentUri, string href, XbrlFragment containingFragment)
         {
@@ -65,7 +66,7 @@ namespace JeffFerguson.Gepsio
             {
                 return new ReferenceLinkbaseDocument(containingDocumentUri, href, containingFragment);
             }
-            throw new NotSupportedException($"Linkbase node has unsupported child node with local name {firstChildLocalName} in document {href} at URI {containingDocumentUri}.");
+            return null;
         }
 
         //------------------------------------------------------------------------------------
@@ -77,26 +78,17 @@ namespace JeffFerguson.Gepsio
             {
                 return LinkbaseDocFilename;
             }
-            int FirstPathSeparator = LinkbaseDocFilename.IndexOf(System.IO.Path.DirectorySeparatorChar);
-            if (FirstPathSeparator == -1)
-            {
-                string DocumentUri = ContainingDocumentUri;
-                int LastPathSeparator = DocumentUri.LastIndexOf(System.IO.Path.DirectorySeparatorChar);
-                if (LastPathSeparator == -1)
-                    LastPathSeparator = DocumentUri.LastIndexOf('/');
-                string DocumentPath = DocumentUri.Substring(0, LastPathSeparator + 1);
+            var DocumentUri = ContainingDocumentUri;
+            int LastPathSeparator = DocumentUri.LastIndexOf(System.IO.Path.DirectorySeparatorChar);
+            if (LastPathSeparator == -1)
+                LastPathSeparator = DocumentUri.LastIndexOf('/');
+            string DocumentPath = DocumentUri.Substring(0, LastPathSeparator + 1);
 
-                // Check for remote linkbases when using local files
+            // Check for remote linkbases when using local files.
 
-                if ((DocumentPath.StartsWith("file:///") == true) && (LinkbaseDocFilename.StartsWith("http://") == true))
-                    return LinkbaseDocFilename;
-
-                FullPath = DocumentPath + LinkbaseDocFilename;
-            }
-            else
-            {
-                throw new NotImplementedException("XbrlSchema.GetFullSchemaPath() code path not implemented.");
-            }
+            if ((DocumentPath.StartsWith("file:///") == true) && (LinkbaseDocFilename.StartsWith("http://") == true))
+                return LinkbaseDocFilename;
+            FullPath = DocumentPath + LinkbaseDocFilename;
             return FullPath;
         }